requirements.txt
pyyaml
ffmpeg-python
tqdm
requests
lxml
xmler
pytz
httplib2
apiclient
o365